DOIO Rev 2 Customization
