+++ keyboards/doio/kb16/info.json
@@ -4,8 +4,12 @@
     "maintainer": "HorrorTroll",
     "usb": {
         "vid": "0xD010",
-        "pid": "0x1601",
-        "force_nkro": true
+        "pid": "0x1601"
+    },
+    "host": {
+        "default": {
+            "nkro": true
+        }
     },
     "diode_direction": "COL2ROW",
     "build": {
