+++ keyboards/doio/kb16/keymaps/default/rules.mk
@@ -1,2 +1,5 @@
 # Encoder enabled
 ENCODER_MAP_ENABLE = yes
+CONSOLE_ENABLE = yes
+DEFERRED_EXEC_ENABLE = yes
+LINK_TIME_OPTIMIZATION_ENABLE = yes
