--- keyboards/doio/kb16/keymaps/personal/rules.mk
@@ -1,6 +1,0 @@
-# Encoder enabled
-ENCODER_MAP_ENABLE = yes
-CONSOLE_ENABLE = yes       
-COMMAND_ENABLE = no
-DEFERRED_EXEC_ENABLE = yes
-LINK_TIME_OPTIMIZATION_ENABLE = yes
